Summary of changes:
Add an index action in the users_controller to list all users that exist within the applicartion

Add an index.html.erb template and filled in the code for the view using the app/views/articles/_article.html.erb partial as a starting template.

Add a link to 'Authors' in the navigation partial which leads to the users index page.