User Authentication using helper methods summary:
Adding a current_user and logged_in? helper methods.

Modified the navigation partial and utilized these helper methods to display the username of the user who signed-in and create a link into their individual profiles.

Modified the users_controller so that when a new user has successfully sign-up, they will be immediately signed-in.

Adding a set_user method and used in a before_action callback to de-clutter repetitive code and can only be use for show, edit, and update action in our users_controller.