Delete user functionality summary:
Add a functionality for users to delete their own accounts using a destroy method in the UsersController.

Add a link to delete account to the user's profile dropdown navigation option.

Utilize a couple of existing methods (as before_action methods) to secure the newly created destroy action at controller level.