Admin User summary:
Generated a migration file to add admin field to users table with a default value of false for all users.

Whether a user is admin or not, can be tested by first getting a user object in the rails console and checking user.admin? from there.

To switch a user to an admin, you can use the following command in the rails console(see notepad)